lib/memcopy: use glibc version
the kernel's memcpy and memmove is very inefficient. But the glibc version is
quite fast, in some cases it is 10 times faster than the kernel version. So I
introduce some memory copy macros and functions of the glibc to improve the
kernel version's performance.

The strategy of the memory functions is:
1. Copy bytes until the destination pointer is aligned.
2. Copy words in unrolled loops.  If the source and destination are not
   aligned in the same way, use word memory operations, but shift and merge
   two read words before writing.
3. Copy the few remaining bytes.

Signed-off-by: Miao Xie <miaox*******>
Signed-off-by: engstk <eng.stk@sapo.pt>

Conflicts:

	lib/Makefile
Signed-off-by: flar2 <asegaert@gmail.com>

Change-Id: I5f4f1e5ef76eb4bf9a2d0a8cf40e70abb9738cd5
Signed-off-by: engstk <eng.stk@sapo.pt>